Ensure address width is correct in Shared BRAM
The address of the shared BRAM gets munged to (attempt to)
influence the endianness of the data when writing words narrower (and
wider?) than 32 bits.  I think the current approach looks problematic
because I would have thought one would just invert the LSbs, but instead
the LSbs are subtracted from 0.  Regardless of whether this "works" as
intended, the problem being fixed in this commit relates to the address
that is presented to the simulation BRAM.  That address was "inverse
munged" to undo this endian "cleverness", but unfortunately the mask
script did not update the address width fields of that address
(un-)munging block so it was always hardcoded to 10 bits.  That's a
problem if more bits are needed (e.g. for 4096 32-bit words).  The fix
is to just uncomment a line that was commented out for some reason
(perhaps an inadvertent remmnant of some ill fated late night debugging
session?).