Add vivadorc file
This file is analagous to startsg in that it looks for a local config
file to get local configuation details and then sets up the environment
for running vivado.  Unlike startsg, this script is A) only for vivado
(i.e. not for MATLAB/Simulink/System Generator) and B) is intended to
allow the user to run vivado from the command line rather than starting
MATLAB/Simulink/System Generator and C) it should be sourced rather than
exectued directly.

Maybe someday (soon?) this file will also include MATLAB stuff and
startsg will source this file.  That's probably a cleaner setup than
having two parallel ways of setting up the environemnt.