Specify async clocks via set_clock_groups command
Vivado will consider the timing requirements for all inter-clock paths
unless it is told that the clocks are asynchronous with each other.
Without telling Vivado about the async clocks, it will deduce some very
strange and difficult to meet timing constraints.  Attempting to meet
these bizarre timing constraints can cause Vivado to fail to meet other
"real" timing constraints.  To avoid this, we use the set_clock_groups
command to inform Vivado of the asynchronous clocks so that it will not
worry about the timing between these clock domains.